[#3] 카테고리 리스트 컴포넌트 이동
