+++ src/components/organisms/CategoriesWrapper.tsx
@@ -1,11 +1,11 @@
-import styled from "styled-components";
-import CategoriesList from "../moecules/CategoriesList";
-import Category from "@/types/category";
+import styled from 'styled-components';
+import CategoriesList from '../CategoriesList/CategoriesList';
+import Category from '@/types/category';
 
 type Props = {
   categories: Category[];
   category?: string;
-}
+};
 
 const CategoriesWrapperBox = styled.nav`
   margin-bottom: 20px;
@@ -15,17 +15,18 @@ const CategoriesWrapperBox = styled.nav`
     font-family: 'Bebas Neue', 'Do Hyeon', cursive;
     font-size: 48px;
   }
-`
+`;
 
-const CategoriesWrapper = ({categories, category} : Props) => {
+const CategoriesWrapper = ({ categories, category }: Props) => {
   return (
     <CategoriesWrapperBox>
-      <h2>
-        Categories
-      </h2>
-      <CategoriesList categories={categories} category={category}/>
+      <h2>Categories</h2>
+      <CategoriesList
+        categories={categories}
+        category={category}
+      />
     </CategoriesWrapperBox>
-  )
-}
+  );
+};
 
 export default CategoriesWrapper;
